feat: integrated kit package in server-kit
