+++ packages/kit/src/index.ts
@@ -7,7 +7,8 @@
 
 export * from './error';
 export * from './event';
-export * from './jwt';
+export * from './json-web-key';
+export * from './json-web-token';
 export * from './oauth2';
 export * from './openid';
 export * from './utils';
