+++ packages/kit/src/json-web-token/constants.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2024.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+export enum JWTAlgorithm {
+    'HS256' = 'HS256',
+    'HS384' = 'HS384',
+    'HS512' = 'HS512',
+    'RS256' = 'RS256',
+    'RS384' = 'RS384',
+    'RS512' = 'RS512',
+    'ES256' = 'ES256',
+    'ES384' = 'ES384',
+    // 'ES512' = 'ES512',
+    'PS256' = 'PS256',
+    'PS384' = 'PS384',
+    'PS512' = 'PS512',
+}
