+++ packages/kit/src/json-web-token/index.ts
@@ -5,4 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './constants';
 export * from './types';
