+++ packages/kit/src/json-web-token/types.ts
