+++ packages/kit/src/oauth2/constants.ts
@@ -5,21 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export enum JWTAlgorithm {
-    'HS256' = 'HS256',
-    'HS384' = 'HS384',
-    'HS512' = 'HS512',
-    'RS256' = 'RS256',
-    'RS384' = 'RS384',
-    'RS512' = 'RS512',
-    'ES256' = 'ES256',
-    'ES384' = 'ES384',
-    // 'ES512' = 'ES512',
-    'PS256' = 'PS256',
-    'PS384' = 'PS384',
-    'PS512' = 'PS512',
-}
-
 export enum OAuth2TokenGrant {
     AUTHORIZATION_CODE = 'authorization_code',
     CLIENT_CREDENTIALS = 'client_credentials',
@@ -38,4 +23,11 @@ export enum OAuth2SubKind {
     CLIENT = 'client',
     USER = 'user',
     ROBOT = 'robot',
+}
+
+export enum OAuth2AuthorizationResponseType {
+    NONE = 'none',
+    CODE = 'code',
+    TOKEN = 'token',
+    ID_TOKEN = 'id_token',
 }
