+++ packages/server-kit/package.json
@@ -62,10 +62,10 @@
         "winston": "^3.13.0"
     },
     "devDependencies": {
-        "@authup/core-kit": "^1.0.0-beta.9"
+        "@authup/kit": "^1.0.0-beta.9"
     },
     "peerDependencies": {
-        "@authup/core-kit": "^1.0.0-beta.9"
+        "@authup/kit": "^1.0.0-beta.9"
     },
     "publishConfig": {
         "access": "public"
