+++ packages/server-kit/src/crypto/json-web-token/type.ts
@@ -5,7 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import type { JWTAlgorithm } from '@authup/core-kit';
+import type { JWTAlgorithm } from '@authup/kit';
 
 export type TokenRSAAlgorithm = `${JWTAlgorithm.RS256}` |
     `${JWTAlgorithm.RS384}` |
