+++ packages/server-kit/src/crypto/key-pair/utils/check.ts
@@ -5,7 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { isObject } from '@authup/core-kit';
+import { isObject } from '@authup/kit';
 import type { KeyPair } from '../type';
 
 export function isKeyPair(data: unknown) : data is KeyPair {
