+++ packages/server-kit/src/domain-event/type.ts
@@ -5,7 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import type { DomainsEventContext } from '@authup/core-kit';
+import type { EventPayload } from '@authup/kit';
 
 export type DomainEventChannelName = string | ((id?: string | number) => string);
 export type DomainEventDestination = {
@@ -15,13 +15,8 @@ export type DomainEventDestination = {
 
 export type DomainEventDestinations = DomainEventDestination[];
 
-export type DomainEventPublishContent = {
-    type: string,
-    event: string,
-    data: Record<string, any>
-};
 export type DomainEventPublishContext<
-    T extends DomainEventPublishContent = DomainsEventContext,
+    T extends EventPayload = EventPayload,
 > = {
     content: T,
     destinations: DomainEventDestinations
